Added raw server response logging for auth internal errors
It seems like a handful of users (including myself) have run into
“internal errors” that are not actually internal errors, but rather
unexpected errors that this SDK does not handle. Instead of just eating
these errors, I’d like to log it so it is easier to track down issues
and so developers can maybe figure out the problem themselves without
having to reach out to support.

I decided to go with the simplest solution, which is just to
JSON.stringify() the raw server response as part of the error message.
However, another option is to add a new property such as
“rawServerResponse” to FirebaseError.errorInfo. I’m interested to hear
what you think is the best option.

Change-Id: I962d8b67d7a1753b271e5c7da411f5d9324567e5